#1016: Use FORCE_SCRIPT_NAME setting on AsgiHandler (#1064)
The 'handle' method on AsgiHandler only veryfies the 'root_path' key of it's
'scope', but it should first verify the FORCE_SCRIPT_NAME parameter from the
project's settings. It defines a prefix that precedes all urls and is used as
the value of the SCRIPT_NAME environment variable.

This parameter is used by the 'runserver' command, but not by the asgi
application, resulting on url's being resolved by django with false
paths.